requirements.txt
streamlitlangchainlangchain-ollamalangchain-chromascikit-learnnumpypandas